+++ lib/tests/Api256Shim.test.ts
@@ -102,6 +102,8 @@ import("../Api256Shim").then((Recrypt) => {
                     const plaintext = api.generatePlaintext();
                     expect(plaintext).to.be.a("Uint8Array");
                     expect(plaintext).to.have.lengthOf(384);
+                    const plaintext2 = api.generatePlaintext();
+                    expect(plaintext).to.not.deep.equal(plaintext2);
                 });
             });
 
@@ -426,6 +428,33 @@ import("../Api256Shim").then((Recrypt) => {
             });
         });
 
+        describe("EncryptedSearch", () => {
+            const encSearch = new Recrypt.EncryptedSearch();
+            describe("generateHashesForString", () => {
+                it("generates valid hashes for string", () => {
+                    const queryResult1 = encSearch.generateHashesForString("ironcore labs", new Uint8Array([1]), "red");
+                    const queryResult2 = encSearch.generateHashesForString("ironcore laps", new Uint8Array([1]), "red");
+                    const queryResult3 = encSearch.generateHashesForString("ironcore labs", new Uint8Array([1]), "bed");
+                    const queryResult4 = encSearch.generateHashesForString("ironcore labs", new Uint8Array([2]), "red");
+
+                    expect(queryResult1).to.be.a("Uint32Array");
+                    expect(queryResult1).to.have.lengthOf(8);
+                    expect(queryResult1).not.to.deep.equal(queryResult2);
+                    expect(queryResult1).not.to.deep.equal(queryResult3);
+                    expect(queryResult1).not.to.deep.equal(queryResult4);
+                });
+            });
+            describe("generateHashesForStringWithPadding", () => {
+                it("generates more hashes than without padding", () => {
+                    const queryResult = encSearch.generateHashesForString("ironcore labs", new Uint8Array([1]), undefined);
+                    const dataResult = encSearch.generateHashesForStringWithPadding("ironcore labs", new Uint8Array([1]), undefined);
+
+                    expect(dataResult).to.have.length.be.above(8);
+                    expect(Array.from(dataResult)).to.include.members(Array.from(queryResult));
+                });
+            });
+        });
+
         describe("transformKeyToBytes256", () => {
             it("converts a transform key into bytes", () => {
                 const fromPrivateKey = api.generateKeyPair().privateKey;
@@ -524,7 +553,6 @@ import("../Api256Shim").then((Recrypt) => {
                 const key1 = new Uint8Array([1, 2, 0, 221, 116, 9, 241, 149, 253, 82, 219, 45, 60, 186, 93, 114, 202, 103, 9, 191, 29, 148, 18, 27, 243, 116, 136, 1, 180, 1, 1, 0]);
                 //prettier-ignore
                 const key2 = new Uint8Array([1, 1, 1, 104, 32, 121, 170, 221, 21, 229, 188, 159, 140, 164, 44, 173, 30, 151, 210, 60, 34, 10, 160, 186, 168, 36, 102, 174, 64, 0, 0, 1]);
-                console.log(Recrypt.subtractPrivateKeys(key1, key2));
                 expect(Recrypt.subtractPrivateKeys(key1, key2)).to.deep.equal(
                     //prettier-ignore
                     new Uint8Array([0, 0, 255, 117, 83, 144, 70, 184, 231, 109, 30, 141, 176, 22, 48, 197, 171, 207, 55, 130, 251, 137, 113, 97, 75, 80, 33, 83, 116, 1, 0, 255])
@@ -544,7 +572,6 @@ import("../Api256Shim").then((Recrypt) => {
                 expect(() => Recrypt.setRandomSeed({} as any)).to.throw;
                 expect(() => Recrypt.setRandomSeed([] as any)).to.throw;
 
-                expect(() => Recrypt.setRandomSeed(new Uint8Array(30))).to.throw;
                 expect(() => Recrypt.setRandomSeed(new Uint8Array(0))).to.throw;
                 expect(() => Recrypt.setRandomSeed(new Uint8Array(31))).to.throw;
             });